+++ packages/string-strip-html/changelog.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] - 2018-07-03
+
+Breaking changes: `opts.dumpLinkHrefsNearby` was previously Boolean. Now it's a plain object and its key `enabled` (`opts.dumpLinkHrefsNearby.enabled`) does the same thing that `opts.dumpLinkHrefsNearby` did before `v3`.
+
+This makes it easier for us to contain all new `opts.dumpLinkHrefsNearby` settings in one place:
+
+```js
+{
+  ignoreTags: [],
+  stripTogetherWithTheirContents: stripTogetherWithTheirContentsDefaults,
+  skipHtmlDecoding: false,
+  returnRangesOnly: false,
+  trimOnlySpaces: false,
+  dumpLinkHrefsNearby: { // <------ CHANGED!
+    enabled: false, // <-------- 💥 NEW!
+    putOnNewLine: false, // <--- 💥 NEW!
+    wrapHeads: "", // <--------- 💥 NEW!
+    wrapTails: "" // <---------- 💥 NEW!
+  }
+}
+```
+
+- ✨ Now, input string is returned trimmed of whitespace in the beginning and in the end.
+
 ## [2.4.0] - 2018-06-20
 
 - ✨ Two `range-` dependencies have been renamed, namely [ranges-push](https://www.npmjs.com/package/ranges-push) and [ranges-apply](https://www.npmjs.com/package/ranges-apply). We tapped them.
@@ -84,3 +108,4 @@ One day I noticed that my Nunjucks code (just a greater-than comparison against 
 [2.2.0]: https://bitbucket.org/codsen/string-strip-html/branches/compare/v2.2.0%0Dv2.1.2#diff
 [2.3.0]: https://bitbucket.org/codsen/string-strip-html/branches/compare/v2.3.0%0Dv2.2.2#diff
 [2.4.0]: https://bitbucket.org/codsen/string-strip-html/branches/compare/v2.4.0%0Dv2.3.0#diff
+[3.0.0]: https://bitbucket.org/codsen/string-strip-html/branches/compare/v3.0.0%0Dv2.4.1#diff